Reactify prototype
