+++ prototype/package.json
@@ -4,14 +4,20 @@
     "main": "index.js",
     "license": "MIT",
     "dependencies": {
-        "jquery": "^3.2.1"
+        "classnames": "^2.2.5",
+        "jquery": "^3.2.1",
+        "react": "^16.2.0",
+        "react-dom": "^16.2.0"
     },
     "scripts": {
-        "start": "parcel src/index.html"
+        "start": "parcel src/index.html",
+        "start-v1": "parcel src/v1.html"
     },
     "devDependencies": {
         "babel-eslint": "^8.2.1",
         "babel-preset-env": "^1.6.1",
+        "babel-preset-react": "^6.24.1",
+        "babel-preset-stage-2": "^6.24.1",
         "eslint": "^4.15.0",
         "eslint-config-prettier": "^2.9.0",
         "eslint-plugin-prettier": "^2.5.0",
