+++ prototype/src/index.html
@@ -6,24 +6,7 @@
         <title>Waterworks Prototype</title>
     </head>
     <body>
-        <div class="illustration">
-            <div class="target"></div>
-            <div class="device"></div>
-        </div>
-        <div class="data">
-            <div class="dimension -alpha">
-                <span class="label">alpha</span>
-                <span class="value"></span>
-            </div>
-            <div class="dimension -beta">
-                <span class="label">beta</span>
-                <span class="value"></span>
-            </div>
-            <div class="dimension -dump">
-                <span class="label">dump</span>
-                <span class="value"></span>
-            </div>
-        </div>
+        <div id="app"></div>
         <script src="js/main.js"></script>
     </body>
 </html>
