+++ prototype/src/js/constants.js
@@ -1,0 +1,21 @@
+export const alphaConstants = {
+    target: 180,
+    slack: 20
+};
+
+export const betaConstants = {
+    target: 68,
+    slack: 13,
+    min: 0,
+    max: 90
+};
+
+export const orienterStates = {
+    start: 'start',
+    betaAligned: 'betaAligned',
+    alphaAligned: 'alphaAligned'
+};
+
+export const cssOffsetPropertyName = '--offset';
+
+export const hideOrienterDelay = 700;
