+++ prototype/src/js/utils.js
@@ -1,0 +1,7 @@
+export function delay(duration, func) {
+    return window.setTimeout(func, duration);
+}
+
+export function cancelDelay(id) {
+    return window.clearTimeout(id);
+}
