+++ prototype/src/sass/_app-root.scss
@@ -1,0 +1,28 @@
+#app {
+    position: absolute;
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: stretch;
+    justify-content: flex-start;
+    width: 100%;
+    height: 100%;
+    min-height: 50vh;
+
+    > .app-root {
+        width: 100%;
+        height: 100%;
+    }
+}
+
+.app-root {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: stretch;
+    justify-content: flex-start;
+
+    > .device-orienter {
+        flex: none;
+        width: 100%;
+        height: 100%;
+    }
+}
