+++ prototype/src/sass/_base.scss
@@ -1,0 +1,24 @@
+@import url('https://fonts.googleapis.com/css?family=Alegreya+Sans:300,400,700');
+
+*,
+*::before,
+*::after {
+    box-sizing: inherit;
+}
+
+html {
+    --offset: 0;
+    box-sizing: border-box;
+    font-size: 62.5%;
+}
+
+body {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: stretch;
+    justify-content: flex-start;
+    width: 100vw;
+    min-height: 80vh;
+    font-family: 'Alegreya Sans', sans-serif;
+    font-weight: 400;
+}
