+++ prototype/src/sass/_device-orienter.scss
@@ -1,0 +1,23 @@
+.device-orienter {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: stretch;
+    justify-content: flex-start;
+    background-color: #efefef;
+
+    > .axis-orienter {
+        flex: none;
+        width: 100%;
+        height: 100%;
+    }
+
+    > .success {
+        display: flex;
+        flex: auto;
+        flex-flow: column nowrap;
+        align-items: center;
+        justify-content: center;
+        text-align: center;
+        font-size: 32rem;
+    }
+}
