+++ prototype/src/v1.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate" />
+        <title>Waterworks Prototype</title>
+    </head>
+    <body>
+        <div class="illustration">
+            <div class="target"></div>
+            <div class="device"></div>
+        </div>
+        <div class="data">
+            <div class="dimension -alpha">
+                <span class="label">alpha</span>
+                <span class="value"></span>
+            </div>
+            <div class="dimension -beta">
+                <span class="label">beta</span>
+                <span class="value"></span>
+            </div>
+            <div class="dimension -dump">
+                <span class="label">dump</span>
+                <span class="value"></span>
+            </div>
+        </div>
+        <script src="js/v1.js"></script>
+    </body>
+</html>
